requirements.txt
flask>=3.0.0
transformers>=4.41.0
accelerate>=0.26.0
bitsandbytes>=0.43.0
Pillow>=10.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
